+++ Jenkinsfile
@@ -1,15 +1,11 @@
 #!/usr/bin/env groovy
 import com.continuousx.jenkins.LogLevelType
-import com.continuousx.jenkins.pipelines.config.PipelineConfigMavenBuild
+import com.continuousx.jenkins.pipelines.mavenbuild.PipelineMavenBuildConfig
 import org.jenkinsci.plugins.workflow.libs.Library
 
 @Library(['jenkins-cx-shared-library@feature/new-stages']) _
 
-PipelineConfigMavenBuild pipelineConfig = new PipelineConfigMavenBuild(
-        logLevelType: LogLevelType.DEBUG,
+PipelineGlobal(new PipelineMavenBuildConfig(logLevelType: LogLevelType.DEBUG)
+        .configStageJenkinsConvertPluginsTxt(true,true)
+        .configStageMavenCompile(true,true)
 )
-
-pipelineConfig.configStageJenkinsConvertPluginsTxt(true,true)
-pipelineConfig.configStageMavenCompile(true,true)
-
-PipelineGlobal(pipelineConfig)
