+++ src/com/continuousx/jenkins/pipelines/Pipeline.groovy
@@ -1,7 +1,5 @@
 package com.continuousx.jenkins.pipelines
 
-import com.continuousx.jenkins.pipelines.config.PipelineConfig
-
 interface Pipeline {
     PipelineConfig getConfig()
 }
