+++ src/com/continuousx/jenkins/pipelines/PipelineConfig.groovy
@@ -1,0 +1,8 @@
+package com.continuousx.jenkins.pipelines
+
+import com.continuousx.jenkins.LogLevelType
+
+interface PipelineConfig extends Serializable {
+    PipelineType getType()
+    LogLevelType getLogLevelType()
+}
