--- src/com/continuousx/jenkins/pipelines/config/PipelineConfig.groovy
@@ -1,9 +1,0 @@
-package com.continuousx.jenkins.pipelines.config
-
-import com.continuousx.jenkins.LogLevelType
-import com.continuousx.jenkins.pipelines.PipelineType
-
-interface PipelineConfig extends Serializable {
-    PipelineType getPipelineType()
-    LogLevelType getLogLevelType()
-}
