--- src/com/continuousx/jenkins/pipelines/config/PipelineConfigMetadata.groovy
@@ -1,10 +1,0 @@
-package com.continuousx.jenkins.pipelines.config
-
-import com.continuousx.jenkins.pipelines.PipelineType
-import groovy.transform.TypeChecked
-
-@TypeChecked
-class PipelineConfigMetadata implements Serializable {
-    String name
-    PipelineType type
-}
