--- src/com/continuousx/jenkins/pipelines/config/PipelineConfigStageConfigEntry.groovy
@@ -1,11 +1,0 @@
-package com.continuousx.jenkins.pipelines.config
-
-import com.continuousx.jenkins.stages.config.StageConfig
-import com.continuousx.jenkins.stages.StageType
-import groovy.transform.TypeChecked
-
-@TypeChecked
-class PipelineConfigStageConfigEntry implements Serializable {
-    StageType name
-    StageConfig stageConfig
-}
