+++ vars/PipelineGlobal.groovy
@@ -1,19 +1,19 @@
-import com.continuousx.jenkins.pipelines.PipelineNextLevel
+import com.continuousx.jenkins.pipelines.nextlevel.PipelineNextLevel
 import com.continuousx.jenkins.pipelines.PipelineType
-import com.continuousx.jenkins.pipelines.config.PipelineConfig
+import com.continuousx.jenkins.pipelines.PipelineConfig
 
 @SuppressWarnings(['GroovyAssignabilityCheck', 'GroovyVariableCanBeFinal'])
 def call(PipelineConfig config) {
 
-    switch(config.getPipelineType()) {
+    switch(config.getType()) {
         case PipelineType.PIPELINE_MAVEN_BUILD:
-            PipelineMavenBuild(new com.continuousx.jenkins.pipelines.PipelineMavenBuild(this, config))
+            PipelineMavenBuild(new com.continuousx.jenkins.pipelines.mavenbuild.PipelineMavenBuild(this, config))
             break
         case PipelineType.PIPELINE_JENKINS_SHARED_LIB:
             PipelineSharedLib(config)
             break
         case PipelineType.PIPELINE_NEXT_LEVEL:
-            log.info "run Pipeline: ${config.getPipelineType()}"
+            log.info "run Pipeline: ${config.getType()}"
             new PipelineNextLevel(this, config).run()
             break
         default:
